backend/requirements.txt
Flask
flask-cors
openai>=1.0.0
python-dotenv
numpy
chromadb
pdfminer.six
gunicorn
google-generativeai
